updating jouranal apis

+++ src/main/java/com/mk/journalApp/service/JournalEntryService.java
@@ -51,11 +51,20 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
-    public void delete(ObjectId id, String username){
-        User user = userService.findByUserName(username);
-        user.getJournalEntries().removeIf(x->x.getId() == id);
-        userService.saveUser(user);
-        journalEntryRepository.deleteById(id);
+    @Transactional
+    public boolean deleteById(ObjectId id, String username) {
+        boolean removed = false;
+        try{
+            User user = userService.findByUserName(username);
+            removed = user.getJournalEntries().removeIf(x->x.getId().equals(id));
+            if(removed){
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+        }
+        catch (Exception e){
+            throw new RuntimeException("An error occurred while deleting the entry ",e);
+        }
+        return removed;
     }
-
 }
